Fix 암호화코드와 유저서비스 기능 분리
단일책임원칙 적용
