+++ src/main/java/com/dev/careers/service/encryption/PasswordEncryption.java
@@ -1,0 +1,8 @@
+package com.dev.careers.service.encryption;
+
+import java.security.NoSuchAlgorithmException;
+
+public interface PasswordEncryption {
+    String makeSalt();
+    String hashing(byte[] password, String salt) throws NoSuchAlgorithmException;
+}
